V4

+++ utils/functions.py
@@ -7,13 +7,18 @@ import asyncio
 import logging
 import random
 import re
-from itertools import zip_longest
+from contextlib import suppress
+from typing import Callable, TYPE_CHECKING, TypeVar
 
 import discord
-from fuzzywuzzy import fuzz, process
+import disnake
+from rapidfuzz import fuzz, process
 
 from cogs5e.models.errors import NoSelectionElements, SelectionCancelled
 from utils import constants, enums
+
+if TYPE_CHECKING:
+    from utils.context import AvraeContext
 
 log = logging.getLogger(__name__)
 sentinel = object()
@@ -40,14 +45,18 @@ def get_positivity(string):
 
 
 # ==== search / select menus ====
-def search(list_to_search: list, value, key, cutoff=5, return_key=False, strict=False):
+_HaystackT = TypeVar("_HaystackT")
+
+
+def search(
+    list_to_search: list[_HaystackT], value: str, key: Callable[[_HaystackT], str], cutoff=5, strict=False
+) -> tuple[_HaystackT | list[_HaystackT], bool]:
     """Fuzzy searches a list for an object
     result can be either an object or list of objects
     :param list_to_search: The list to search.
     :param value: The value to search for.
     :param key: A function defining what to search for.
     :param cutoff: The scorer cutoff value for fuzzy searching.
-    :param return_key: Whether to return the key of the object that matched or the object itself.
     :param strict: If True, will only search for exact matches.
     :returns: A two-tuple (result, strict)"""
     # there is nothing to search
@@ -82,102 +91,30 @@ def search(list_to_search: list, value, key, cutoff=5, return_key=False, strict=
         results = exact_matches
 
     if len(results) > 1:
-        if return_key:
-            return [key(r) for r in results], False
-        else:
-            return results, False
+        return results, False
     elif not results:
         return [], False
     else:
-        if return_key:
-            return key(results[0]), True
-        else:
-            return results[0], True
+        return results[0], True
 
 
-async def search_and_select(
+def paginate(choices: list[_HaystackT], per_page: int) -> list[list[_HaystackT]]:
+    out = []
+    for start_idx in range(0, len(choices), per_page):
+        out.append(choices[start_idx : start_idx + per_page])
+    return out
+
+
+async def get_selection(
     ctx,
-    list_to_search: list,
-    query,
-    key,
-    cutoff=5,
-    return_key=False,
+    choices: list[_HaystackT],
+    key: Callable[[_HaystackT], str],
+    delete=True,
     pm=False,
     message=None,
-    list_filter=None,
-    selectkey=None,
-    search_func=search,
-    return_metadata=False,
-    strip_query_quotes=True,
+    force_select=False,
 ):
-    """
-    Searches a list for an object matching the key, and prompts user to select on multiple matches.
-    Guaranteed to return a result - raises if there is no result.
-
-    :param ctx: The context of the search.
-    :param list_to_search: The list of objects to search.
-    :param query: The value to search for.
-    :param key: How to search - compares key(obj) to value
-    :param cutoff: The cutoff percentage of fuzzy searches.
-    :param return_key: Whether to return key(match) or match.
-    :param pm: Whether to PM the user the select prompt.
-    :param message: A message to add to the select prompt.
-    :param list_filter: A filter to filter the list to search by.
-    :param selectkey: If supplied, each option will display as selectkey(opt) in the select prompt.
-    :param search_func: The function to use to search.
-    :param return_metadata: Whether to return a metadata object {num_options, chosen_index}.
-    :param strip_query_quotes: Whether to strip quotes from the query.
-    """
-    if list_filter:
-        list_to_search = list(filter(list_filter, list_to_search))
-
-    if search_func is None:
-        search_func = search
-
-    if strip_query_quotes:
-        query = query.strip("\"'")
-
-    if asyncio.iscoroutinefunction(search_func):
-        result = await search_func(list_to_search, query, key, cutoff, return_key)
-    else:
-        result = search_func(list_to_search, query, key, cutoff, return_key)
-
-    if result is None:
-        raise NoSelectionElements("No matches found.")
-    strict = result[1]
-    results = result[0]
-
-    if strict:
-        result = results
-    else:
-        if len(results) == 0:
-            raise NoSelectionElements()
-
-        first_result = results[0]
-        confidence = fuzz.partial_ratio(key(first_result).lower(), query.lower())
-        if len(results) == 1 and confidence > 75:
-            result = first_result
-        else:
-            if selectkey:
-                options = [(selectkey(r), r) for r in results]
-            elif return_key:
-                options = [(r, r) for r in results]
-            else:
-                options = [(key(r), r) for r in results]
-            result = await get_selection(ctx, options, pm=pm, message=message, force_select=True)
-    if not return_metadata:
-        return result
-    metadata = {"num_options": 1 if strict else len(results), "chosen_index": 0 if strict else results.index(result)}
-    return result, metadata
-
-
-def paginate(iterable, n, fillvalue=None):
-    args = [iter(iterable)] * n
-    return [i for i in zip_longest(*args, fillvalue=fillvalue) if i is not None]
-
-
-async def get_selection(ctx, choices, delete=True, pm=False, message=None, force_select=False):
-    """Returns the selected choice, or None. Choices should be a list of two-tuples of (name, choice).
+    """Returns the selected choice, or raises an error.
     If delete is True, will delete the selection message and the response.
     If length of choices is 1, will return the only choice unless force_select is True.
 
@@ -186,39 +123,41 @@ async def get_selection(ctx, choices, delete=True, pm=False, message=None, force
     if len(choices) == 0:
         raise NoSelectionElements()
     elif len(choices) == 1 and not force_select:
-        return choices[0][1]
+        return choices[0]
 
     page = 0
     pages = paginate(choices, 10)
     m = None
-    selectMsg = None
+    select_msg = None
 
     def chk(msg):
-        valid = [str(v) for v in range(1, len(choices) + 1)] + ["c", "n", "p"]
-        return msg.author == ctx.author and msg.channel == ctx.channel and msg.content.lower() in valid
+        content = msg.content.lower()
+        valid = content in ("c", "n", "p")
+        try:
+            valid = valid or (1 <= int(content) <= (len(choices) + 1))
+        except ValueError:
+            pass
+        return msg.author == ctx.author and msg.channel == ctx.channel and valid
 
     for n in range(200):
         _choices = pages[page]
-        names = [o[0] for o in _choices if o]
+        names = [key(o) for o in _choices]
         embed = discord.Embed()
         embed.title = "Multiple Matches Found"
-        selectStr = 'Which one were you looking for? (Type the number or "c" to cancel)\n'
+        select_str = "Which one were you looking for? (Type the number or `c` to cancel)\n"
         if len(pages) > 1:
-            selectStr += "`n` to go to the next page, or `p` for previous\n"
+            select_str += "`n` to go to the next page, or `p` for previous\n"
             embed.set_footer(text=f"Page {page + 1}/{len(pages)}")
         for i, r in enumerate(names):
-            selectStr += f"**[{i + 1 + page * 10}]** - {r}\n"
-        embed.description = selectStr
+            select_str += f"**[{i + 1 + page * 10}]** - {r}\n"
+        embed.description = select_str
         embed.colour = random.randint(0, 0xFFFFFF)
         if message:
             embed.add_field(name="Note", value=message, inline=False)
-        if selectMsg:
-            try:
-                await selectMsg.delete()
-            except:
-                pass
+        if select_msg:
+            await try_delete(select_msg)
         if not pm:
-            selectMsg = await ctx.channel.send(embed=embed)
+            select_msg = await ctx.channel.send(embed=embed)
         else:
             embed.add_field(
                 name="Instructions",
@@ -226,7 +165,7 @@ async def get_selection(ctx, choices, delete=True, pm=False, message=None, force
                 "you to hide the monster name.",
                 inline=False,
             )
-            selectMsg = await ctx.author.send(embed=embed)
+            select_msg = await ctx.author.send(embed=embed)
 
         try:
             m = await ctx.bot.wait_for("message", timeout=30, check=chk)
@@ -249,14 +188,74 @@ async def get_selection(ctx, choices, delete=True, pm=False, message=None, force
             break
 
     if delete and not pm:
-        try:
-            await selectMsg.delete()
-            await m.delete()
-        except:
-            pass
+        with suppress(disnake.HTTPException):
+            await select_msg.delete()
+            if m is not None:
+                await m.delete()
     if m is None or m.content.lower() == "c":
         raise SelectionCancelled()
-    return choices[int(m.content) - 1][1]
+    return choices[int(m.content) - 1]
+
+
+async def search_and_select(
+    ctx: "AvraeContext",
+    list_to_search: list[_HaystackT],
+    query: str,
+    key: Callable[[_HaystackT], str],
+    cutoff=5,
+    pm=False,
+    message=None,
+    list_filter=None,
+    selectkey=None,
+    return_metadata=False,
+    strip_query_quotes=True,
+    selector=get_selection,
+) -> _HaystackT:
+    """
+    Searches a list for an object matching the key, and prompts user to select on multiple matches.
+    Guaranteed to return a result - raises if there is no result.
+
+    :param ctx: The context of the search.
+    :param list_to_search: The list of objects to search.
+    :param query: The value to search for.
+    :param key: How to search - compares key(obj) to value
+    :param cutoff: The cutoff percentage of fuzzy searches.
+    :param pm: Whether to PM the user the select prompt.
+    :param message: A message to add to the select prompt.
+    :param list_filter: A filter to filter the list to search by.
+    :param selectkey: If supplied, each option will display as selectkey(opt) in the select prompt.
+    :param return_metadata: Whether to return a metadata object {num_options, chosen_index}.
+    :param strip_query_quotes: Whether to strip quotes from the query.
+    :param selector: The coroutine to use to select a result if multiple results are possible.
+    """
+    if list_filter:
+        list_to_search = list(filter(list_filter, list_to_search))
+
+    if strip_query_quotes:
+        query = query.strip("\"'")
+
+    result = search(list_to_search, query, key, cutoff)
+
+    if result is None:
+        raise NoSelectionElements("No matches found.")
+    results, strict = result
+
+    if strict:
+        result = results
+    else:
+        if len(results) == 0:
+            raise NoSelectionElements()
+
+        first_result = results[0]
+        confidence = fuzz.partial_ratio(key(first_result).lower(), query.lower())
+        if len(results) == 1 and confidence > 75:
+            result = first_result
+        else:
+            result = await selector(ctx, results, key=selectkey or key, pm=pm, message=message, force_select=True)
+    if not return_metadata:
+        return result
+    metadata = {"num_options": 1 if strict else len(results), "chosen_index": 0 if strict else results.index(result)}
+    return result, metadata
 
 
 async def confirm(ctx, message, delete_msgs=False, response_check=get_positivity):
@@ -478,13 +477,13 @@ async def get_guild_member(guild, member_id):
     return None
 
 
-def reconcile_adv(adv=False, dis=False, ea=False) -> enums.AdvantageType:
+def reconcile_adv(adv=False, dis=False, eadv=False) -> enums.AdvantageType:
     """
     Reconciles sets of advantage passed in
 
     :param adv: Combined advantage
     :param dis: Combined disadvantage
-    :param ea:  Combined elven accuracy
+    :param eadv:  Combined elven accuracy
     :return: The combined advantage result
     """
     result = 0
@@ -492,7 +491,7 @@ def reconcile_adv(adv=False, dis=False, ea=False) -> enums.AdvantageType:
         result += 1
     if dis:
         result += -1
-    if ea and not dis:
+    if eadv and not dis:
         return enums.AdvantageType.ELVEN
     return enums.AdvantageType(result)
 
@@ -511,3 +510,21 @@ def exactly_one(iterable):
     if next(iterable, sentinel) is sentinel:
         return retval
     return None
+
+
+split_regex = re.compile(r"\W+")
+
+
+def get_initials(name: str) -> str:
+    """
+    Returns the initials for a given string if its multiple words, otherwise returns the first two letters in uppercase.
+    """
+    # Mainly used for monster combatants in initiative
+    split_name = split_regex.split(name)
+
+    # Single word, return first two letters, uppercase
+    if len(split_name) == 1:
+        return name[:2].upper()
+
+    # Multiple words, find initials
+    return "".join(word[0] for word in split_name if word)
